Windows generation
Generate + build for Windows. Other platforms coming soon.

Decision was made here to use the CSP inside the _deps_ folder as the thing that we'll write our interface includes against, rather than duplicating/submoduling the CSP include dir.

This means that you'll need to run configure if you want to browse the CSP library from inside this repo.

It strikes me that keeping what is wrapped/what isn't wrapped in sync may be painful. I mean, it already _is_ painful but there's opportunities to be quite informative about it with this setup. Something to think on.