Wrap LogSystem
Wrap LogSystem. Only one test for now, more coming soon. (Not really logsystem tests, but mechanism tests)

This shows that we can wrap CSP interfaces without needing to modify CSP, adding an adaptive pattern that injects the callback in via std::function capture.
This does mean we need to care about callback lifetimes, which I _think_ is currently managed fine and dandy via the static lifetime of the callbacks in the SWIG dispatch (modified in our fork), but not sure. Simple enough to add a manager for this if we need to at any rate.