Remove hope-based GC tests
Im a little surprised the native interop is so difficult to make deterministic in failure cases from the C# side, I thought there'd be a special managed exception or something like that. :shrug:

We _can_ make the interfaces resistant to this, but it involves making sure unmanaged calls to collected managed objects can never happen, which involves pinning the managed adapters.

This is a decision to be made, because you can leave this in control of the developers, or in control of the generated interop layer. There's appeals to both.