Add Csp_List typemap
Adapted from std_vector.i, taken from SWIGS own typemaps. Is still 90% that code. I wouldn't recommend doing a full style review of this as I don't think we should be overly interested in rewriting this file to meet our style tastes, best to keep is as matching the example as possible imo.

Maintains all the patterns from std_vector as much as possible, but adapting to the interfaces that csp_common_list has.