Remove stripping from further utilities types
We need the `being/end` iterators for the typemap, which are not exported. This is an easy way to maintain them for use in the interface.

We'll probably need to do the same for all the interop types, but have only done the two i'm certain we need right now.