Add List tests
Add tests for List wrapping behaviour. Note the ValueList test. We'll need to add at least one more test for reference list behaviour when we expose one, although ideally we'd just do a data drive test and run all the tests on both modes.

As the specific case for this list exposure is a List property on LoginState, add a property test for that as well.