Search Csharp .i files first.
Interesting, this matters on non-windows platform. I suspect it's todo with compiler (ie, clang throws a warning based on intrinsics)

We are including std_except.i, but there's a version in both the standard and csharp lib dirs. The standard one objects to being directly included, so make sure we get the csharp one.