Add Map->Dictionary tests
Same as List, same as it ever was.

I generated a large chunk of the Dict tests using an LLM. Not sure how I feel about it. They're good by my reading but ... I dunno. Makes me uncomfortable how easy it was to get them all spat out. I suspect I spent the same amount of energy verifying them as I would have writing them, and yet I feel like I'm less familiar with them than I would have been otherwise.

It's like it shifts the effort to the back of the work, when normally the energy to break the blank-page is the hard part.