[OPE-2975] feat: Add equatable types
Add a way to have value types in the SWIG interface.

This is done by implementing the IEquatable interface, which requires equality and hash operators in the C++ code.

Have added minimal implementations in the CSP lib to Vector2/3/4 just to test this

Discussion with the C# team led to the decision that it is most expectable for containers to remain reference comparable. Therefore some tests around that that have been removed.